Added Shooting code for the Player
Added projectile.png (from Google Drive) into the Sprites folder,
sliced them. Made a prefab for firePoint (the point where the
projectile is fired from) which is a child of Player. Made a prefab for
the projectile. Made ShooterController and ProjectileController scripts
for shooting projectiles. One problem I’ll fix later is that if you
fire a projectile and then immediately walk into it, it’ll disappear.