Number of shots limited, changed speed of projectile
I made the speed of the mixtape a bit faster. The projectile destroys
itself after half a second of being fired. Limited the number of times
a projectile can be fired to 2 (can’t have more than 2 mixtapes on the
screen at one time) so that you can’t spam the Return key as much.