Player Health and Boss Health
I added the necessary elements to display health in both the player and
the boss.

I added a player health script and a boss health script.  I also added
health bars to each, which are displayed using the UI slider.  I also
made it so when the player’s projectile hits the boss, it does damage
to the boss.  When the boss touches the player, the player gets damaged.

If the player dies, he gets transported to level 1.  If the boss dies,
the player gets transported to level 2.