Confidence Meter
I added the confidence meter!

I made it so it is visible to the screen using UI Slider like before.

I made it so you can kill the chicken in one hit.  When you kill the
chicken, it adds confidence.  When the chicken touches the player, it
subtracts confidence.

The confidence is fixed in the boss battle.