Made A Big Chicken
This is just another alternative Enemy I call Big Chicken. It's a Big
Chicken. So like a chicken... but bigger. It just patrols for right now
and doesn't flip. But, when the player is close enough it shoots an Egg.
For the Egg I made it a new prefab called nme_Projectile, it damages the
player as is destroyed on collision with player or player projectile.
Some fire rate balancing is probably needed.