Confidence has an impact on Player Damage and Boss Damage
I have made it so the confidence actually does something useful.

When the player has high confidence, they do high damage to the boss.
However, the boss now does high damage to them.  As expect, players who
are experts at this game will have high confidence and be able to
tackle the challenge while dealing high damage.

When the player has low confidence, they do a low amount of damage, but
the boss will also deal a low amount of damage.  This is for when
players are not good at the game.  The boss will have a hard time
killing them, so they will have plenty of practice to shoot down the
boss with their low damage.