Boss2 v1.02
Commented the bejeebers out of it

Addforce just not digging it. So just so you guys can see it using
transform.velocity

The boss will chase when health is high enough, and simply stop and hope
you go away for healing times.

Destroy the bombers and he will go rebuild them.