Drunk V1.01
K proper commit this time. This is the prototype for Drunk.

The drunk alternates between throwing and lobbing bottles at the player
depending on distance.

Also has a rage ability where he "charges up" then fires a bunch of
bottles really quickly at the player. The Rage ability has a cooldown
and a short duration.

The code is all written but he is not working. I don't know why but it
seems like the decision tree is just not doing anything at the moment as
none of his variables are increasing with the time and his debug logs
aren't showing up. I'm trying to figure out why DecisionTree isn't
attaching or working.