Bug Fixes and Balancing
I fixed all the Barrel/Tree/Lamp issues.  They now have walls that are
not labeled as platform.

There STILL is the case where the player can be stuck in the “not
grounded” state if an enemy pushes them against a wall.  However I
fixed this by making a timer.  If the player is not grounded for more
than 2 seconds, they will then be grounded again.  The reason it is 2
seconds is to prevent double jumping from being in the air for too long.

I fixed some balancing issues.  The game is now much easier.  All
bosses are definitely beatable even with full confidence.