Shoot through clouds and big chicken eggs
Changed each Cloud’s Layer to “Cloud” in every Level. Made a change to
the ProjectileController script so that the player’s projectile ignores
collisions with GameObjects that have a Cloud layer.

Also changed the layer of the Big Chicken egg (nme_projectile prefab)
to “Boss Projectile” so that collisions between it and the player’s
projectiles are ignored. (You were able to destroy the egg by shooting
it before).