Added pause menu to Level 1,2,3
Made a pause menu prefab + script; added the prefab to level 1,2,3. The
pause menu can either resume game or quit to main menu.

The Esc key is used to pause the game.

Also added the Start Screen scene to File -> Build Settings to get rid
of that “Scene _____ couldn’t be loaded….” error.